refactor(string-adder): refactor package structure
