+++ src/main/kotlin/adder/Main.kt
@@ -1,11 +1,15 @@
 package adder
 
 import adder.model.Numbers
+import adder.view.InputView
+import adder.view.ResultView
 
 fun main() {
     InputView.printQuestion()
     var input = readLine()
+
     while (input == null) {
+        InputView.printQuestion()
         input = readLine()
     }
