+++ src/main/kotlin/adder/view/ResultView.kt
@@ -1,4 +1,4 @@
-package adder
+package adder.view
 
 object ResultView {
     fun printSum(sum: Int) {
